test: move time-picker i18n tests to a separate file
The `custom functions` suite grew to cover both `i18n` functions and the
cases where the object is reassigned. It moves to `i18n.test.js`, split
into `parseTime`, `formatTime` and `reassigned` suites, so that new cases
have a place to go.

No test is added, removed or changed.
